+++ api_yamdb/api/urls.py
@@ -1,9 +1,14 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
 
-from .views import (
-UserCreation, JWTTokenConfirmation, UserViewSet,
-CategoryViewSet, GenreViewSet, TitleViewSet)
+from api.views import (UserCreation,
+                       JWTTokenConfirmation,
+                       UserViewSet,
+                       CategoryViewSet,
+                       CommentViewSet,
+                       GenreViewSet,
+                       ReviewViewSet,
+                       TitleViewSet)
 
 router_v1 = DefaultRouter()
 
@@ -11,13 +16,21 @@ router_v1.register(r'users', UserViewSet, basename='users')
 router_v1.register(r'categories', CategoryViewSet, basename='category')
 router_v1.register(r'genres', GenreViewSet, basename='genre')
 router_v1.register(r'titles', TitleViewSet, basename='title')
+router_v1.register(
+    r'titles/(?P<title_id>\d+)/reviews',
+    ReviewViewSet,
+    basename='review'
+)
+router_v1.register(
+    r'titles/(?P<title_id>\d+)/reviews/(?P<review_id>\d+)/comments',
+    CommentViewSet,
+    basename='comment'
+)
 
-urlpatterns= [
+urlpatterns = [
     path('v1/', include(router_v1.urls)),
     path('v1/auth/signup/', UserCreation.as_view(),
          name='signup'),
     path('v1/auth/token/', JWTTokenConfirmation.as_view(),
          name='get_token'),
-     path('', include('ratings.urls')),
 ]
-
