--- api_yamdb/ratings/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-
-from ratings.models import Comment, Review
-
-admin.site.register(Comment)
-admin.site.register(Review)
