--- api_yamdb/ratings/urls.py
@@ -1,16 +1,0 @@
-from django.urls import path, include
-from rest_framework.routers import DefaultRouter
-
-from ratings.views import (
-    ReviewViewSet,
-    CommentViewSet,
-)
-
-router_v1 = DefaultRouter()
-
-router_v1.register(r'titles/(?P<title_id>\d+)/reviews', ReviewViewSet, basename='reviews')
-router_v1.register(r'titles/(?P<title_id>\d+)/reviews/(?P<review_id>\d+)/comments', CommentViewSet, basename='comments')
-
-urlpatterns= [
-    path('v1/', include(router_v1.urls)),
-]
