+++ api_yamdb/reviews/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import User, Title, Genre, Category
+from .models import User, Title, Genre, Category, Comment, Review
 
 
 class UserAdmin(admin.ModelAdmin):
@@ -16,7 +16,10 @@ class UserAdmin(admin.ModelAdmin):
     search_fields = ('username',)
     empty_value_display = '-пусто-'
 
+
 admin.site.register(User, UserAdmin)
 admin.site.register(Title)
 admin.site.register(Genre)
 admin.site.register(Category)
+admin.site.register(Comment)
+admin.site.register(Review)
