+++ api_yamdb/reviews/models.py
@@ -1,9 +1,10 @@
 from django.conf import settings
 from django.contrib.auth.models import AbstractUser
 from django.db import models
-from .validators import valid_year 
-from api.validators import validate_username, validate_regex_username
+from django.core.validators import MaxValueValidator, MinValueValidator
 
+from .validators import valid_year
+from api.validators import validate_username, validate_regex_username
 
 
 class Category(models.Model):
@@ -55,9 +56,11 @@ class Title(models.Model):
         blank=True,
         null=True,
         verbose_name='Описание произведения',)
-    genre = models.ManyToManyField(
+    genre = models.ForeignKey(
         Genre,
         blank=False,
+        null=True,
+        on_delete=models.SET_NULL,
         verbose_name='Жанр произведения',)
     category = models.ForeignKey(
         Category,
@@ -78,7 +81,7 @@ class User(AbstractUser):
         unique=True,
         validators=(
             validate_username,
-            validate_regex_username ),
+            validate_regex_username),
     )
     email = models.EmailField(
         'Email',
@@ -113,5 +116,82 @@ class User(AbstractUser):
 
     def __str__(self):
         return 'Пользователь - {}'.format(self.username)
+
     class Meta:
         verbose_name_plural = 'Пользователи'
+
+
+class Review(models.Model):
+    """Модель отзыва."""
+    title = models.ForeignKey(
+        Title,
+        on_delete=models.CASCADE,
+        related_name='reviews',
+        verbose_name='ID произведения',
+    )
+    author = models.OneToOneField(
+        User,
+        on_delete=models.CASCADE,
+        related_name='reviews',
+        verbose_name='Автор отзыва',
+    )
+    text = models.TextField(
+        blank=False,
+        null=False,
+        verbose_name='Текст отзыва',
+    )
+    score = models.IntegerField(
+        blank=False,
+        null=False,
+        verbose_name='Оценка',
+        validators=[
+            MaxValueValidator(10),
+            MinValueValidator(1)
+        ]
+    )
+    pub_date = models.DateTimeField(
+        verbose_name='Дата публикации',
+        auto_now_add=True
+    )
+
+    class Meta:
+        verbose_name = 'Review'
+        verbose_name_plural = 'Reviews'
+        ordering = ['pub_date']
+
+    def __str__(self):
+        return self.text[:30]
+
+
+class Comment(models.Model):
+    review = models.ForeignKey(
+        Review,
+        blank=False,
+        null=False,
+        on_delete=models.CASCADE,
+        related_name='comments',
+        verbose_name='Комментарий',
+    )
+    author = models.ForeignKey(
+        User,
+        on_delete=models.CASCADE,
+        related_name='comments',
+        verbose_name='Автор комментария',
+    )
+    text = models.TextField(
+        blank=False,
+        null=False,
+        verbose_name='Текст комментария',
+    )
+    pub_date = models.DateTimeField(
+        verbose_name='Дата публикации',
+        auto_now_add=True
+    )
+
+    class Meta:
+        verbose_name = 'Comment'
+        verbose_name_plural = 'Comments'
+        ordering = ['pub_date']
+
+    def __str__(self):
+        return self.text[:15]
